automatically backed up by learn

+++ lib/cash_register.rb
@@ -11,6 +11,6 @@ class CashRegister
   end
 
   def apply_discount
-    self.total = self.total.to_f - self.total.to_f * self.discount.to_f
+    self.total = self.total.to_i - self.total.to_i * self.discount.to_i
   end
 end
